requirements.txt
fastapi
pymongo
dnspython
python-dotenv
google-genai